+++ shop-master/src/main/java/com/shop/dto/ItemFormDto.java
@@ -27,9 +27,9 @@ public class ItemFormDto {
     private String itemSNm;
 
     @NotNull(message = "가격은 필수 입력 값입니다.")
-    private Integer ori_price;
+    private double ori_price;
     
-    private Integer percent;
+    private double percent;
 
     @NotBlank(message = "상품 상세는 필수 입력 값입니다.")
     private String itemDetail;
@@ -61,7 +61,7 @@ public class ItemFormDto {
         return modelMapper.map(item,ItemFormDto.class);
     }
     
-    public int getPrice() {
+    public double getPrice() {
     	return ori_price - (ori_price *percent/100);
     }
