+++ shop-master/src/main/java/com/shop/dto/OrderItemDto.java
@@ -17,7 +17,7 @@ public class OrderItemDto {
     private String itemNm; //상품명
     private int count; //주문 수량
 
-    private int orderPrice; //주문 금액
+    private double orderPrice; //주문 금액
     private String imgUrl; //상품 이미지 경로
 
 }
