+++ shop-master/src/main/java/com/shop/entity/Order.java
@@ -49,8 +49,8 @@ public class Order extends BaseEntity {
         return order;
     }
 
-    public int getTotalPrice() {
-        int totalPrice = 0;
+    public double getTotalPrice() {
+        double totalPrice = 0;
         for(OrderItem orderItem : orderItems){
             totalPrice += orderItem.getTotalPrice();
         }
